+++ realm/realm-library/src/objectServer/java/io/realm/internal/SyncObjectServerFacade.java
@@ -39,8 +39,8 @@ import io.realm.exceptions.RealmException;
 import io.realm.internal.android.AndroidCapabilities;
 import io.realm.internal.network.NetworkStateReceiver;
 import io.realm.internal.objectstore.OsAsyncOpenTask;
-import io.realm.internal.sync.permissions.ObjectPermissionsModule;
 import io.realm.sync.Subscription;
+import io.realm.internal.sync.permissions.ObjectPermissionsModule;
 
 @SuppressWarnings({"unused", "WeakerAccess"}) // Used through reflection. See ObjectServerFacade
 @Keep